src/requirements.txt
docker
flask
bcrypt
pyjwt
sqlalchemy
alembic
pymysql